+++ include/UHI/NativeImGuiHost.h
@@ -3,9 +3,10 @@
 namespace UHI::NativeImGuiHost
 {
     using RenderCallback = void (*)();
+    using CloseCallback = void (*)();
 
     bool InstallRendererHook();
-    bool Register(RenderCallback callback);
+    bool Register(RenderCallback callback, CloseCallback closed);
     bool Open();
     bool Toggle();
     bool Close();